Criação do fluxo completo de gestão de mensalidades, com rotas para geração e pagamento manual. A entidade Cliente foi refatorada para calcular seu status dinamicamente com base no histórico de faturas. Também foi implementada uma rotina automática (@Scheduled) para gerar as cobranças todo dia 1º e a tipagem financeira foi atualizada para BigDecimal, garantindo maior precisão.

+++ src/main/java/ClientControl/DTO/ResumoDTO.java
@@ -2,15 +2,15 @@ package ClientControl.DTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import java.math.BigDecimal;
 
 @Getter
 @AllArgsConstructor
 public class ResumoDTO {
-
     private long totalClientes;
-    private double totalPago;
-    private double totalPendente;
-    private double totalGeral;
-    private double totalVencido;
-    private double totalFaltaReceber;
+    private BigDecimal totalPago;
+    private BigDecimal totalPendente;
+    private BigDecimal totalGeral;
+    private BigDecimal totalVencido;
+    private BigDecimal totalFaltaReceber;
 }
